moving schemas outside of components
