+++ webapp/app/modules/categories/components/forms/child/create.tsx
@@ -26,7 +26,7 @@ import {
 import { Input } from "~/shared/components/ui/input";
 import { Textarea } from "~/shared/components/ui/textarea";
 import { Icon } from "~/shared/types/icon";
-import { schema } from "../../schemas/child/create";
+import { schema } from "../../../schemas/child/create";
 
 interface Child {
   name: string
