+++ webapp/app/modules/categories/components/forms/child/update.tsx
@@ -26,7 +26,7 @@ import {
 import { Input } from "~/shared/components/ui/input";
 import { Textarea } from "~/shared/components/ui/textarea";
 import { Icon } from "~/shared/types/icon";
-import { schema } from "../../schemas/child/update";
+import { schema } from "../../../schemas/child/update";
 
 interface Child {
   id: number
